requirements.txt
tensorflow-gpu==2.3.1
numpy
opencv-python==3.4.4.19
pandas
scipy
tqdm
matplotlib
pillow
pyqt5
sklearn
scikit-image
dlib==19.6.1
jupyter
